+++ taxsystem/api/taxsystem/admin.py
@@ -15,7 +15,7 @@ from django.utils.translation import gettext_lazy as _
 # AA TaxSystem
 from taxsystem.api.helpers import get_manage_corporation
 from taxsystem.api.taxsystem.helpers.administration import _delete_member
-from taxsystem.api.taxsystem.helpers.payments import _payments_actions
+from taxsystem.api.taxsystem.helpers.filters import _filter_actions
 from taxsystem.api.taxsystem.helpers.paymentsystem import (
     _payment_system_actions,
     _payments_info,
@@ -25,8 +25,9 @@ from taxsystem.api.taxsystem.helpers.statistics import (
     _get_statistics_dict,
 )
 from taxsystem.helpers import lazy
+from taxsystem.models.filters import JournalFilter
 from taxsystem.models.logs import AdminLogs
-from taxsystem.models.tax import Members, Payments, PaymentSystem
+from taxsystem.models.tax import Members, PaymentSystem
 from taxsystem.models.wallet import (
     CorporationWalletDivision,
     CorporationWalletJournalEntry,
@@ -234,13 +235,11 @@ class AdminApiEndpoints:
             return output
 
         @api.get(
-            "corporation/{corporation_id}/character/{character_id}/view/payments/",
+            "corporation/{corporation_id}/filter_set/{filter_set_id}/view/filter/",
             response={200: list, 403: str, 404: str},
             tags=self.tags,
         )
-        def get_main_character_payments(
-            request, corporation_id: int, character_id: int
-        ):
+        def get_filter_set_filters(request, corporation_id: int, filter_set_id: int):
             owner, perms = get_manage_corporation(request, corporation_id)
 
             if owner is None:
@@ -249,64 +248,36 @@ class AdminApiEndpoints:
             if perms is False:
                 return 403, "Permission Denied"
 
-            payments = Payments.objects.filter(
-                account__owner=owner,
-                account__user__profile__main_character__character_id=character_id,
+            filters = JournalFilter.objects.filter(
+                filter_set__pk=filter_set_id,
             )
 
-            if not payments:
-                return 404, "No Payments Found"
+            output = []
 
-            # Create a dict for the character
-            payments_char_dict = {
-                "title": "Payments for",
-                "character_id": character_id,
-                "character_portrait": lazy.get_character_portrait_url(
-                    character_id, size=32, as_html=True
-                ),
-                "character_name": payments[0].account.name,
-            }
+            for filter_obj in filters:
+                if filter_obj.filter_type == JournalFilter.FilterType.AMOUNT:
+                    value = f"{intcomma(filter_obj.value)} ISK"
+                else:
+                    value = filter_obj.value
 
-            # Create a dict for each payment
-            payments_dict = {}
-            for payment in payments:
-                try:
-                    character_id = (
-                        payment.account.user.profile.main_character.character_id
-                    )
-                    portrait = lazy.get_character_portrait_url(
-                        character_id, size=32, as_html=True
-                    )
-                except AttributeError:
-                    portrait = ""
-
-                actions = _payments_actions(corporation_id, payment, perms, request)
-                amount = f"{intcomma(payment.amount, use_l10n=True)} ISK"
-
-                payments_dict[payment.pk] = {
-                    "payment_id": payment.pk,
-                    "character_portrait": portrait,
-                    "character_name": payment.account.name,
-                    "payment_date": payment.formatted_payment_date,
-                    "amount": amount,
-                    "request_status": payment.get_request_status_display(),
-                    "reviser": payment.reviser,
-                    "division": payment.division,
-                    "reason": payment.reason,
-                    "actions": actions,
+                filter_dict = {
+                    "id": filter_obj.pk,
+                    "filter_set": filter_obj.filter_set,
+                    "filter_type": filter_obj.get_filter_type_display(),
+                    "value": value,
+                    "actions": _filter_actions(
+                        corporation_id=corporation_id,
+                        filter_obj=filter_obj,
+                        perms=perms,
+                        request=request,
+                    ),
                 }
-
-            # Add payments to the character dict
-            payments_char_dict["payments"] = payments_dict
-
-            context = {
-                "entity_pk": corporation_id,
-                "entity_type": "character",
-                "character": payments_char_dict,
-            }
+                output.append(filter_dict)
 
             return render(
                 request,
-                "taxsystem/modals/view_character_payments.html",
-                context,
+                "taxsystem/modals/view_filter.html",
+                context={
+                    "filters": output,
+                },
             )