+++ taxsystem/apps.py
@@ -14,8 +14,3 @@ class TaxSystemConfig(AppConfig):
     name = "taxsystem"
     label = "taxsystem"
     verbose_name = f"Tax System v{__version__}"
-
-    def ready(self):
-        # pylint: disable=import-outside-toplevel,unused-import
-        # AA TaxSystem
-        import taxsystem.signals